requirements.txt
pyopenssl
datetime
socket